# back/docs/openapi.yaml
openapi: 3.0.0
paths:
...
  /consumables:
    post:
      operationId: ConsumablesController_create
      parameters: []
      requestBody:
        required: true
        description: New consumable payload
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: Potion de soin
                consumableTypeId:
                  type: number
                  example: 1
                quantity:
                  type: number
                  example: 10
                cost:
                  type: number
                  example: 25
              required:
                - name
                - consumableTypeId
                - quantity
                - cost
              additionalProperties: false
      responses:
        '201':
          description: Consumable created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: number
                    example: 1
                  name:
                    type: string
                    example: Potion de soin
                  consumableTypeId:
                    type: number
                    example: 1
                  quantity:
                    type: number
                    example: 10
                  cost:
                    type: number
                    example: 25
                  createdAt:
                    type: string
                    format: date-time
                    example: '2025-12-12T10:00:00.000Z'
                  updatedAt:
                    type: string
                    format: date-time
                    example: '2025-12-12T10:00:00.000Z'
      security:
        - bearer: []
      tags:
        - Consumables
    get:
      operationId: ConsumablesController_findAll
      parameters: []
      responses:
        '200':
          description: List of consumables
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: number
                      example: 1
                    name:
                      type: string
                      example: Potion de soin
                    consumableTypeId:
                      type: number
                      example: 1
                    quantity:
                      type: number
                      example: 10
                    cost:
                      type: number
                      example: 25
                    createdAt:
                      type: string
                      format: date-time
                      example: '2025-12-12T10:00:00.000Z'
                    updatedAt:
                      type: string
                      format: date-time
                      example: '2025-12-12T10:00:00.000Z'
      security:
        - bearer: []
      tags:
        - Consumables
  /consumables/{id}:
    get:
      operationId: ConsumablesController_findOne
      parameters:
        - name: id
          required: true
          in: path
          description: Consumable ID
          schema:
            example: 1
            type: number
      responses:
        '200':
          description: Consumable by id
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: number
                    example: 1
                  name:
                    type: string
                    example: Potion de soin
                  consumableTypeId:
                    type: number
                    example: 1
                  quantity:
                    type: number
                    example: 10
                  cost:
                    type: number
                    example: 25
                  createdAt:
                    type: string
                    format: date-time
                    example: '2025-12-12T10:00:00.000Z'
                  updatedAt:
                    type: string
                    format: date-time
                    example: '2025-12-12T10:00:00.000Z'
      security:
        - bearer: []
      tags:
        - Consumables
    patch:
      operationId: ConsumablesController_update
      parameters:
        - name: id
          required: true
          in: path
          description: Consumable ID
          schema:
            example: 1
            type: number
      requestBody:
        required: true
        description: Fields to update (partial)
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: Grande potion de soin
                consumableTypeId:
                  type: number
                  example: 2
                quantity:
                  type: number
                  example: 15
                cost:
                  type: number
                  example: 50
              additionalProperties: false
      responses:
        '200':
          description: Updated consumable
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: number
                    example: 1
                  name:
                    type: string
                    example: Grande potion de soin
                  consumableTypeId:
                    type: number
                    example: 2
                  quantity:
                    type: number
                    example: 15
                  cost:
                    type: number
                    example: 50
                  createdAt:
                    type: string
                    format: date-time
                    example: '2025-12-12T10:00:00.000Z'
                  updatedAt:
                    type: string
                    format: date-time
                    example: '2025-12-12T10:30:00.000Z'
      security:
        - bearer: []
      tags:
        - Consumables
    delete:
      operationId: ConsumablesController_remove
      parameters:
        - name: id
          required: true
          in: path
          description: Consumable ID
          schema:
            example: 1
            type: number
      responses:
        '200':
          description: Delete result
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: number
                    example: 1
                  deleted:
                    type: boolean
                    example: true
      security:
        - bearer: []
      tags:
        - Consumables
  /consumables/{id}/purchase:
    post:
      operationId: ConsumablesController_purchase
      parameters:
        - name: id
          required: true
          in: path
          description: Consumable ID
          description: Consumable ID
          schema:
            example: 1
            type: string
      requestBody:
        required: true
        description: Quantité à acheter
        description: Quantité à acheter
        content:
          application/json:
            schema:
              type: object
              properties:
                quantity:
                  type: number
                  example: 3
                  description: Nombre de consommables à acheter
              required:
                - quantity
              type: object
              properties:
                quantity:
                  type: number
                  example: 3
                  description: Nombre de consommables à acheter
              required:
                - quantity
      responses:
        '200':
          description: Achat effectué avec succès
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                example:
                  id: 1
                  name: Potion de soin
                  stock: 7
                  updatedAt: '2025-12-12T10:00:00.000Z'
      tags:
        - Consumables
...
info:
  title: API "Les kips AH"
  description: |-
    This is the API information regarding the project "Les capuches d'Opale".

    C'est un outil de gestion d'aventurier pour une guilde.
  version: 1.0.0
  contact: {}
  termsOfService: https://swagger.io/terms/
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http